backend/pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "lab-management-backend"
version = "0.1.0"
description = "Backend for the laboratory presence and attendance MVP."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.116,<1.0",
  "itsdangerous>=2.2,<3.0",
  "filelock>=3.16,<4.0",
  "pydantic-settings>=2.10,<3.0",
  "pwdlib[argon2]>=0.2,<1.0",
  "python-dotenv>=1.1,<2.0",
  "python-multipart>=0.0.20,<1.0",
  "uvicorn[standard]>=0.35,<1.0",
]

[project.optional-dependencies]
dev = [
  "httpx>=0.28,<1.0",
  "pytest>=8.4,<9.0",
  "ruff>=0.13,<1.0",
]

[tool.setuptools.packages.find]
include = ["app*"]

[tool.pytest.ini_options]
pythonpath = ["."]
testpaths = ["tests"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "B"]
